Changed resizing code and navbar,, change index
Reworked resizing code and navbar, bug fixes, changed index

Common
resizing code has been moved into a single file that is now
shared by
all simulators, reducing the amount of repeated code by a
large amount.

CSS for the navbar is now in its own file which is used by every page to
reduce repeated code.

Fixed a bug in relative velocity where the closest distance was
being
reported as negative While this is correct in the maths behind
the
simulation it would be confusing to the user. The closest distance
is
now always displayed as positive.

Index now has the same navigation
bar as every other page.